nPOPuk 2.11 patch 1 - various bugfixes

+++ tags/npopuk2.11/src/MailBox.c
@@ -32,6 +32,7 @@ HMENU vMenuDone = NULL;
 /* Global Variables */
 extern OPTION op;
 extern TCHAR *DataDir;
+extern int g_soc;
 
 #ifdef _WIN32_WCE_LAGENDA
 extern HMENU hMainMenu;
@@ -46,13 +47,14 @@ extern MAILBOX *MailBox;
 extern int MailBoxCnt;
 extern ADDRESSBOOK *AddressBook;
 
-extern int SelBox, vSelBox;
+extern int SelBox, vSelBox, RecvBox;
 extern int LvSortFlag;
 extern BOOL EndThreadSortFlag;
 extern BOOL PPCFlag;
 
 /* Local Function Prototypes */
-void mailbox_swap_files(HWND hWnd, int i, int j);
+static BOOL mailbox_name_clash(TCHAR *name);
+static void mailbox_swap_files(HWND hWnd, int i, int j);
 
 /*
  * mailbox_init - メールボックスの初期化
@@ -74,18 +76,24 @@ BOOL mailbox_init(void)
 }
 
 /*
- * mailbox_create - メールボックスの追加
+ * mailbox_create - add mailbox(es) to the list
  */
 int mailbox_create(HWND hWnd, int Add, int Index, BOOL ShowFlag, BOOL SelFlag)
 {
 	MAILBOX *TmpMailBox;
 	int count, i;
 
-	//It adds to the list of the mailbox
+	count = MailBoxCnt + Add;
 	if (Index == -1) {
 		Index = MailBoxCnt;
 	}
-	count = MailBoxCnt + Add;
+
+	if (Index < 0 || Index > MailBoxCnt) {
+		TCHAR buf[BUF_SIZE];
+		wsprintf(buf, TEXT("Cannot add mailbox at postion %d"), Index);
+		ErrorMessage(hWnd, buf);
+		return -1;
+	}
 
 	TmpMailBox = (MAILBOX *)mem_calloc(sizeof(MAILBOX) * count);
 	if(TmpMailBox == NULL){
@@ -111,6 +119,10 @@ int mailbox_create(HWND hWnd, int Add, int Index, BOOL ShowFlag, BOOL SelFlag)
 	mem_free(&MailBox);
 	MailBox = TmpMailBox;
 	MailBoxCnt = count;
+
+	if (g_soc != -1 && Index <= RecvBox) {
+		RecvBox++;
+	}
 
 	// rename files above Index, watching out for collisions
 	for (i = MailBoxCnt-1; Add == 1 && i >= Index; i--) {
@@ -166,32 +178,31 @@ int mailbox_create(HWND hWnd, int Add, int Index, BOOL ShowFlag, BOOL SelFlag)
 }
 
 /*
- * mailbox_delete - メールボックスの削除
+ * mailbox_delete - delete mailbox from list
  */
 int mailbox_delete(HWND hWnd, int DelIndex, BOOL CheckFilt, BOOL Select)
 {
 	MAILBOX *TmpMailBox;
 	TCHAR name1[BUF_SIZE], name2[BUF_SIZE];
-	int cnt;
+	int cnt = MailBoxCnt - 1;
 	int i, j;
 
-	//From list of mailbox deletion
-	cnt = MailBoxCnt - 1;
-	TmpMailBox = (MAILBOX *)mem_calloc(sizeof(MAILBOX) * cnt);
-	if(TmpMailBox == NULL){
+	if (DelIndex <= 0 || DelIndex >= MailBoxCnt) {
+		wsprintf(name1, TEXT("Cannot delete mailbox %d"), DelIndex);
+		ErrorMessage(hWnd, name1);
 		return -1;
 	}
 
 	if (DelIndex == vSelBox && hViewWnd != NULL) {
 		SendMessage(hViewWnd, WM_CLOSE, 0, 0);
 	}
+	if (g_soc != -1 && DelIndex <= RecvBox) {
+		RecvBox--;
+	}
 
-	if (DelIndex > 0) {
-		CopyMemory(TmpMailBox, MailBox, DelIndex*sizeof(MAILBOX));
-	}
-	if (DelIndex < cnt) {
-		CopyMemory((TmpMailBox+DelIndex), (MailBox+DelIndex+1), (cnt - DelIndex)*sizeof(MAILBOX));
-	}
+	// if alloc fails, will re-use current block of memory
+	TmpMailBox = (MAILBOX *)mem_calloc(sizeof(MAILBOX) * cnt);
+
 	if ((MailBox+DelIndex)->Filename == NULL)  {
 		TCHAR path[2*BUF_SIZE];
 		wsprintf(name1, TEXT("MailBox%d.dat"), DelIndex - MAILBOX_USER);
@@ -211,9 +222,27 @@ int mailbox_delete(HWND hWnd, int DelIndex, BOOL CheckFilt, BOOL Select)
 	}
 	mailbox_free(MailBox + DelIndex);
 
+	if (TmpMailBox == NULL) {
+		// failed to allocate new, so just use existing block of memory
+		if (DelIndex < cnt) {
+			CopyMemory((MailBox+DelIndex), (MailBox+DelIndex+1), (cnt - DelIndex)*sizeof(MAILBOX));
+		}
+		//(MailBox + cnt) = NULL; or ZeroMemory((MailBox + cnt), sizeof(MAILBOX)); ??
+	} else {
+		if (DelIndex > 0) {
+			CopyMemory(TmpMailBox, MailBox, DelIndex*sizeof(MAILBOX));
+		}
+		if (DelIndex < cnt) {
+			CopyMemory((TmpMailBox+DelIndex), (MailBox+DelIndex+1), (cnt - DelIndex)*sizeof(MAILBOX));
+		}
+		mem_free(&MailBox);
+		MailBox = TmpMailBox;
+	}
+	MailBoxCnt = cnt;
+
 	// rename MailBox%d files above DelIndex, rather than loading&writing them
 	for (i = DelIndex; i < cnt; i++) {
-		if ((TmpMailBox + i)->Filename == NULL) {
+		if ((MailBox + i)->Filename == NULL) {
 			BOOL clash;
 			int k = i;
 			wsprintf(name1, TEXT("MailBox%d.dat"), i + 1 - MAILBOX_USER);
@@ -221,8 +250,8 @@ int mailbox_delete(HWND hWnd, int DelIndex, BOOL CheckFilt, BOOL Select)
 				clash = FALSE;
 				wsprintf(name2, TEXT("MailBox%d.dat"), k - MAILBOX_USER);
 				for (j = 0; j < cnt; j++) {
-					if ((TmpMailBox + j)->Filename != NULL 
-						&& lstrcmpi(name2, (TmpMailBox + j)->Filename) == 0) {
+					if ((MailBox + j)->Filename != NULL 
+						&& lstrcmpi(name2, (MailBox + j)->Filename) == 0) {
 						clash = TRUE;
 						if (k == i) {
 							k = cnt+1;
@@ -234,14 +263,11 @@ int mailbox_delete(HWND hWnd, int DelIndex, BOOL CheckFilt, BOOL Select)
 				}
 			} while (clash == TRUE);
 			if (k != i) {
-				(TmpMailBox + i)->Filename = alloc_copy_t(name2);
+				(MailBox + i)->Filename = alloc_copy_t(name2);
 			}
 			file_rename(hWnd, name1, name2);
 		}
 	}
-	mem_free(&MailBox);
-	MailBox = TmpMailBox;
-	MailBoxCnt = cnt;
 
 	// check that no filter was pointing to this deleted mailbox
 	if (CheckFilt == TRUE) {
@@ -342,16 +368,48 @@ int mailbox_load_now(HWND hWnd, int num, BOOL ask, BOOL do_saveboxes)
 }
 
 /*
+ * mailbox_name_clash - check for Filename=name
+ */
+static BOOL mailbox_name_clash(TCHAR *name)
+{
+	int k;
+	for (k = MAILBOX_USER; k < MailBoxCnt; k++) {
+		if ((MailBox+k) != NULL && (MailBox+k)->Filename != NULL
+			&& lstrcmpi(name, (MailBox+k)->Filename) == 0) {
+			return TRUE;
+		}
+	}
+	return FALSE;
+}
+
+/*
  * mailbox_swap_files - exchange mailbox files (if Filename==NULL)
  */
 void mailbox_swap_files(HWND hWnd, int i, int j)
 {
-	TCHAR name1[BUF_SIZE], name2[BUF_SIZE];
+	TCHAR name1[BUF_SIZE], name2[BUF_SIZE], path[2*BUF_SIZE];
+	int len;
+
 	if ((MailBox+i)->Filename != NULL && (MailBox+j)->Filename != NULL) {
 		return;
 	}
 	if ((MailBox+i)->Filename == NULL) {
 		wsprintf(name1, TEXT("MailBox%d.dat"), i - MAILBOX_USER);
+		str_join_t(path, DataDir, name1, (TCHAR *)-1);
+		len = file_get_size(path);
+		if (len == -1) {
+			// mailbox file "name1" doesn't exist
+			if ((MailBox+j)->Filename == NULL) {
+				wsprintf(name2, TEXT("MailBox%d.dat"), j - MAILBOX_USER);
+				if (mailbox_name_clash(name1)) {
+					// some other mailbox has Filename==name1 -- but it's empty??
+					(MailBox+j)->Filename = alloc_copy_t(name2);
+				} else {
+					file_rename(hWnd, name2, name1);
+				}
+			}
+			return;
+		}
 		if ((MailBox+j)->Filename == NULL) {
 			TCHAR *tmp_name = TEXT("$npop_tmp_mailbox.dat");
 			file_delete(hWnd, tmp_name);
@@ -362,21 +420,18 @@ void mailbox_swap_files(HWND hWnd, int i, int j)
 				(MailBox+i)->Filename = alloc_copy_t(name1);
 				(MailBox+j)->Filename = alloc_copy_t(name2);
 			} else {
-				// these shouldn't fail if the first rename succeeded ...
+				// this could fail if name2 doesn't exist, but that's OK
 				file_rename(hWnd, name2, name1);
+				// this shouldn't fail ...
 				file_rename(hWnd, tmp_name, name2);
 			}
 		} else {
+			// MailBoxi becomes MailBoxj (unless MailBoxj.dat is in use)
 			BOOL found = FALSE;
-			int k;
-			// MailBoxi becomes MailBoxj (unless MailBoxj.dat exists)
 			wsprintf(name2, TEXT("MailBox%d.dat"), j - MAILBOX_USER);
-			for (k = MAILBOX_USER; k < MailBoxCnt; k++) {
-				if ((MailBox+k) != NULL && (MailBox+k)->Filename != NULL
-					&& lstrcmpi(name2, (MailBox+k)->Filename) == 0) {
-					found = TRUE;
-					break;
-				}
+			found = mailbox_name_clash(name2);
+			if (found == FALSE) {
+				file_delete(hWnd, name2);
 			}
 			if (found || file_rename(hWnd, name1, name2) == FALSE) {
 				(MailBox+i)->Filename = alloc_copy_t(name1);
@@ -626,9 +681,9 @@ BOOL mailbox_menu_rebuild(HWND hWnd, BOOL IsAttach) {
 #endif // _WIN32_WCE_PPC
 #else
 			InsertMenu(vMenu, 0, MF_BYPOSITION | MF_POPUP | MF_STRING,
-				(UINT)vCOPYFLY, STR_LIST_MENU_COPYSBOX);
+				(UINT)vCOPYFLY, STR_LIST_MENU_COPYSBOX1);
 			InsertMenu(vMenu, 1, MF_BYPOSITION | MF_POPUP | MF_STRING,
-				(UINT)vMOVEFLY, STR_LIST_MENU_MOVESBOX);
+				(UINT)vMOVEFLY, STR_LIST_MENU_MOVESBOX1);
 #endif // _WIN32_WCE
 			vMenuDone = vMenu;
 		}
@@ -733,7 +788,7 @@ void mailbox_select(HWND hWnd, int Sel)
 	hMenu = GetSubMenu(GetMenu(MainWnd), MailMenuPos);
 #endif
 
-	//of menu Item of the mark is deleted to reply and one for reception the
+	// Delete entries and rebuild appropriate to SendBox or not
 	DeleteMenu(hMenu, ID_MENUITEM_REMESSEGE, MF_BYCOMMAND);
 	DeleteMenu(hMenu, ID_MENUITEM_ALLREMESSEGE, MF_BYCOMMAND);
 	DeleteMenu(hMenu, ID_MENUITEM_DOWNMARK, MF_BYCOMMAND);
@@ -791,7 +846,6 @@ void mailbox_select(HWND hWnd, int Sel)
 			InsertMenu(hMenu, ID_MENUITEM_UNMARK, MF_STRING,
 				ID_MENUITEM_DELMARK, STR_LIST_MENU_DELMARK);
 		}
-
 #endif
 		InsertMenu(hMenu, ID_MENUITEM_DELETE, MF_STRING,
 			ID_MENUITEM_DELATTACH, STR_LIST_MENU_DELATTACH);
